Fix getPower function in the built in type for redstone block entity returning an int rather than a double.

+++ src/main/java/com/workert/robotics/content/computers/ioblocks/redstonedetector/RedstoneDetectorBlockEntity.java
@@ -18,7 +18,7 @@ public class RedstoneDetectorBlockEntity extends SyncedTileEntity implements IOB
 
 	public static RoboScriptClass roboScriptBlockClass = IORoboScriptBlockHelper.createClass()
 			.addMethod("getPower", 0,
-					(vm, fun) -> ((RedstoneDetectorBlockEntity) IORoboScriptBlockHelper.getBlockEntityFromMethod(
+					(vm, fun) -> (double) ((RedstoneDetectorBlockEntity) IORoboScriptBlockHelper.getBlockEntityFromMethod(
 							fun)).redstoneLevel)
 			.build();
 	int redstoneLevel = 0;
